refactor: 로그인 redirectUri 얻어오는 로직 분리
